+++ packages/admin/package.js
@@ -1,5 +1,5 @@
 Package.describe({
-  version: "1.3.7",
+  version: "1.4.0",
   name: 'keplerjs:admin',
   summary: 'keplerjs plugin for administration of platform',
   git: "https://github.com/Keplerjs/Kepler.git"