+++ packages/core/Kepler.js
@@ -2,7 +2,7 @@
 	global container
 */
 Kepler = K = {
-	version: "1.3.7",
+	version: "1.4.0",
 
 	//core models
 	Place: {},       //class for places                  (client)